Move Service Client tests
