+++ packages/safe-service-client/tests/e2e/getServiceInfo.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
-import SafeServiceClient from '../src'
-import { getServiceClient } from './utils'
+import SafeServiceClient from '../../src'
+import { getServiceClient } from '../utils/setupServiceClient'
 
 let serviceSdk: SafeServiceClient
